Time: 165 ms (13.60%) | Memory: 97 MB (6.13%) - LeetSync

+++ 2049-eliminate-maximum-number-of-monsters/eliminate-maximum-number-of-monsters.cpp
@@ -8,7 +8,7 @@ public:
         }
         sort(vec.begin(),vec.end());
         if(vec[0].first==0)return 0;
-        double time=1;
+        int  time=1;
         int cnt=1;
         for(int i=1;i<vec.size();i++){
             if(vec[i].first>time){
